+++ docs/CHANGELOG.md
@@ -9,7 +9,7 @@
 - **Breaking change** Move `deccp2stdhexcp`, `char2codepoint` and `chars2codepoints` from `Properties` to `Utils`
 - Add blocks & planes [#43](https://github.com/noraj/unisec/issues/43)
   - Lib: `Unisec::Blocks` & `Unisec::Planes`
-  - CLI: `unisec blocks list`, `unisec blocks search`, `unisec planes list`, `unisec planes search`
+  - CLI: `unisec blocks list`, `unisec blocks search`, `unisec blocks invalid`, `unisec planes list`, `unisec planes search`
 - New utils method:
   - `Unisec::Utils::String.to_range`
   - `Unisec::Utils::Range.range2codepoint_range`